+++ src/Path/System.php
@@ -4,18 +4,4 @@ namespace Gvozdb\Dumper\Path;
 
 class System extends AbstractPath
 {
-    /**
-     * @var array $config
-     */
-    protected $config = [];
-
-    /**
-     * @param array $config
-     *
-     * @throws \Exception
-     */
-    public function __construct(array $config = [])
-    {
-        parent::__construct($config);
-    }
 }