+++ CHANGELOG.md
@@ -1,25 +1,41 @@
 # Changelog
 
+## [0.2.0] - 2020-07-14
+
+* Refactoring
+* Added output of error messages
+* Added information messages output
+* Added recording of logs to file
+* Added sending of logs to mail
+* Added sending of logs to Telegram Chat
+* Changed README.md
+
+
 ## [0.1.5] - 2019-05-21
 
 * Added param **users** for **Backup::run()**
 
+
 ## [0.1.4] - 2018-11-21
 
-* Added **class_exists** for storage in Dumper\Backup
+* Added **class_exists** for storage in `Dumper\Backup`
 * Changed README.md
+
 
 ## [0.1.3] - 2018-11-07
 
 * Fixed composer.json, changed version of `arhitector/yandex` to ~2.0
 
+
 ## [0.1.2] - 2018-11-07
 
 * Fixed composer.json, added property `prefer-stable`
 
+
 ## [0.1.1] - 2018-11-07
 
 * Changed README.md
+
 
 ## [0.1.0] - 2018-11-06